requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.12.3
    # via yfinance
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
frozendict==2.4.6
    # via yfinance
html5lib==1.1
    # via yfinance
idna==3.10
    # via requests
lxml==5.3.0
    # via yfinance
multitasking==0.0.11
    # via yfinance
mysql-connector-python==9.1.0
    # via -r requirements.in
numpy==2.1.3
    # via
    #   -r requirements.in
    #   pandas
    #   scipy
    #   yfinance
pandas==2.2.3
    # via
    #   -r requirements.in
    #   yfinance
peewee==3.17.8
    # via yfinance
platformdirs==4.3.6
    # via yfinance
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via
    #   pandas
    #   yfinance
requests==2.32.3
    # via yfinance
scipy==1.14.1
    # via -r requirements.in
six==1.16.0
    # via
    #   html5lib
    #   python-dateutil
soupsieve==2.6
    # via beautifulsoup4
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
webencodings==0.5.1
    # via html5lib
yfinance==0.2.50
    # via -r requirements.in
